Added environment variables
The scraper is now configurable with the following environment variables:
* PORT
The port that the scraper runs on in the docker.
* MU_SPARQL_ENDPOINT
The localion of the triple store.
* MU_APPLICATION_GRAPH
The graph to write/read to/from.
* MINIMUM_TIME_FOR_RESCRAPING
The time in secods that needs to have passed before we add a page again to the triple
store. The new page will have it's own UUID and file. If this should never be
done then this variable should be set to -1.