Changed scraper back to vanilla puppeteer
I need to continue investigating this. The headless chrome crawler failed on the
example page at
https://publicatie.gelinkt-notuleren.demo.lblod.info/b93af251-f6e2-11e8-988b-57ef1e8527da/notulen.

Reverting back to vanilla puppeteer and then manually setting the await until
there is no more network activity (instead of the page loaded option on headless
chrome crawler) seemed to do the trick.

I do expect that it would be meaningful to check if hcc can be re-added as that
library had several extra features for dealing with SPA's.